Finalizando...

+++ src/main/java/com/BookStoreV1/BookStore/Service/PublisherService.java
@@ -1,11 +1,18 @@
 package com.BookStoreV1.BookStore.Service;
 
 import com.BookStoreV1.BookStore.Dto.PublisherDTO;
+import com.BookStoreV1.BookStore.Model.Book;
+import com.BookStoreV1.BookStore.Model.Rent;
+import com.BookStoreV1.BookStore.Repository.BookRepository;
+import com.BookStoreV1.BookStore.Repository.RentRepository;
 import com.BookStoreV1.BookStore.Validation.Publisher.PublisherAlreadyExistsException;
 import com.BookStoreV1.BookStore.Validation.Publisher.PublisherNotFoundException;
 import com.BookStoreV1.BookStore.Mapper.PublisherMapper;
 import com.BookStoreV1.BookStore.Model.Publisher;
 import com.BookStoreV1.BookStore.Repository.PublisherRepository;
+import com.BookStoreV1.BookStore.Validation.Publisher.RelatedPublisherRent;
+import com.BookStoreV1.BookStore.Validation.User.RelatedUserRent;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,9 +25,11 @@ public class PublisherService {
     private final static PublisherMapper publisherMapper = PublisherMapper.INSTANCE;
     @Autowired
     private PublisherRepository publisherRepository;
+    @Autowired
+    private BookRepository bookRepository;
 
     public PublisherDTO create(PublisherDTO publisherDTO){
-        verifyExists(publisherDTO.getNome());
+        verifyExists(publisherDTO.getName());
         Publisher publisherToCreate = publisherMapper.toModel(publisherDTO);
         Publisher createdPublisher = publisherRepository.save(publisherToCreate);
         return publisherMapper.toDTO(createdPublisher);
@@ -39,12 +48,22 @@ public class PublisherService {
                 .collect(Collectors.toList());
     }
 
+    @SneakyThrows
     public void delete(Long id){
-        verifyGetIfExists(id);
+        Publisher foundPublisherDelete = verifyGetIfExists(id);
+        if (hasRentsForPublisher(id)) {
+            throw new RelatedPublisherRent(foundPublisherDelete.getId());
+        }
         publisherRepository.deleteById(id);
     }
 
+    public boolean hasRentsForPublisher (Long id) {
+        List<Book> books = bookRepository.findByPublisherId(id);
+        return !books.isEmpty();
+    }
+
     public PublisherDTO update(Long id, PublisherDTO updatedPublisherDTO) {
+        verifyExists(updatedPublisherDTO.getName());
         Publisher foundPublisher = verifyGetIfExists(id);
         updatedPublisherDTO.setId(foundPublisher.getId());
         Publisher publisherUpdate = publisherMapper.toModel(updatedPublisherDTO);
@@ -54,7 +73,7 @@ public class PublisherService {
 
     private void verifyExists(String name) {
         Optional<Publisher> duplicatedPublisher = publisherRepository
-                .findByNome(name);
+                .findByName(name);
         if (duplicatedPublisher.isPresent()){
             throw new PublisherAlreadyExistsException(name);
         }
